# backend/src/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "街口公益捐款 API",
    "description": "街口公益捐款列表 MVP",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    },
                    "ts": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "status",
                    "ts"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/charities": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "ORG",
                "CAMPAIGN",
                "MERCHANDISE"
              ]
            },
            "in": "query",
            "name": "category",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "category_code",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "tab": {
                            "type": "string",
                            "enum": [
                              "ORG",
                              "CAMPAIGN",
                              "MERCHANDISE"
                            ]
                          },
                          "category_codes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "logo_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "amount_raised": {
                            "type": "integer"
                          },
                          "amount_goal": {
                            "type": "integer",
                            "nullable": true
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "banner_image_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "org_name": {
                            "type": "string",
                            "nullable": true
                          },
                          "tags": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "product_image_url": {
                            "type": "string",
                            "nullable": true
                          },
                          "price_ntd": {
                            "type": "integer",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "description",
                          "tab",
                          "category_codes",
                          "logo_url",
                          "amount_raised",
                          "amount_goal",
                          "created_at",
                          "banner_image_url",
                          "org_name",
                          "tags",
                          "product_image_url",
                          "price_ntd"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "next_cursor": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "items",
                    "next_cursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "invalid"
                      ]
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message",
                          "code"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/charities/{id}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "tab": {
                      "type": "string",
                      "enum": [
                        "ORG",
                        "CAMPAIGN",
                        "MERCHANDISE"
                      ]
                    },
                    "category_codes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "logo_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "amount_raised": {
                      "type": "integer"
                    },
                    "amount_goal": {
                      "type": "integer",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "banner_image_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "org_name": {
                      "type": "string",
                      "nullable": true
                    },
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "product_image_url": {
                      "type": "string",
                      "nullable": true
                    },
                    "price_ntd": {
                      "type": "integer",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "tab",
                    "category_codes",
                    "logo_url",
                    "amount_raised",
                    "amount_goal",
                    "created_at",
                    "banner_image_url",
                    "org_name",
                    "tags",
                    "product_image_url",
                    "price_ntd"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "enum": [
                        "not_found"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/categories": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "code": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "code",
                          "label"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "items"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}